+++ tst/RegenerateSnapshotTest.cpp
@@ -41,12 +41,51 @@ BOOST_AUTO_TEST_SUITE(regenerate_tests);
                 fs::path("output") / fs::path(snapshotName + ".regen.esp"),
                 0
         );
-        generator.generateFromExisting(fs::path("regenerate_snapshots") / fs::path(snapshotName) / fs::path("input.esp"));
+        RegenerateOptions options;
+        generator.generateFromExisting(
+                fs::path("regenerate_snapshots") / fs::path(snapshotName) / fs::path("input.esp"),
+                options
+        );
 
         BOOST_TEST_CONTEXT("Snapshot test for: " + snapshotName) {
             verifyEspsMatch(
                     fs::path("output") / fs::path(snapshotName + ".regen.esp"),
                     fs::path("regenerate_snapshots") / fs::path(snapshotName) / fs::path("expected.esp")
+            );
+        }
+    }
+
+    BOOST_DATA_TEST_CASE(
+            RegenWithBase,
+            boost::unit_test::data::make(getSnapshotDirectories()),
+            snapshotName
+    ) {
+        auto base = fs::path("regenerate_snapshots") / fs::path("UpdatedBase.esp");
+        auto input = fs::path("regenerate_snapshots") / fs::path(snapshotName) / fs::path("input.esp");
+        auto output = fs::path("output") / fs::path(snapshotName + ".regenbase.esp");
+
+        Generator generator(
+                makeNullLogger(),
+                [](int, const std::string&) {},
+                [](const std::string&) {},
+                [](const std::exception& err) { throw err; },
+                fs::path("regenerate_snapshots") / fs::path(snapshotName) / fs::path("configuration.ini"),
+                std::vector<fs::path>({base}),
+                output,
+                0
+        );
+
+        RegenerateOptions options;
+        options.basePlugins.insert(base.filename().string());
+        generator.generateFromExisting(
+                input,
+                options
+        );
+
+        BOOST_TEST_CONTEXT("Snapshot test for: " + snapshotName) {
+            verifyEspsMatch(
+                    input,
+                    output
             );
         }
     }